backend/requirements.txt
fastapi 
uvicorn
sqlalchemy 
asyncpg 
langchain 
python-dotenv
httpx
langchain
langchain-google-genai